} Changes in nfet and pfet
The line c_pc.add_ref(poly_con) is commented out to prevent the contacts from being drawn. The following values are changed: con_comp_enc = 0.075 con_pl_enc = 0.08 This applies to both the nfet and pfet functions. The name of c_pc is changed to poly con1.

In the viastack() function, m_enc is changed to 0.08.